+++ Changes.md
@@ -14,6 +14,7 @@ Improvements
 - Dispatcher : Simplified jobs by removing tasks for nodes - such as Wedge - that do no work of their own. This is particularly noticeable in TractorDispatcher, resulting in simpler job graphs in the Tractor dashboard. This behaviour is enabled
 by default but can be temporarily disabled by setting the `GAFFERDISPATCH_OMIT_EMPTY_TASKS` environment variable to a value of `0`. In future, the environment variable will be removed.
 - TractorDispatcher : Added `startPaused` plug.
+- TractorDispatcher, LocalDispatcher : Added context variable summary to task names.
 
 Fixes
 -----
@@ -27,6 +28,7 @@ API
 ---
 
 - MetadataAlgo : Added `firstViewableAncestor()` and `firstViewableNode()` functions.
+- Dispatcher : Added `name()` method to TaskBatch.
 
 1.6.13.0 (relative to 1.6.12.0)
 ========